+++ packages/adk-cli/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @iqai/adk-cli
 
+## 0.3.28
+
+### Patch Changes
+
+- 8d5ba1e: ADK WEB now supports voice input
+- Updated dependencies [8d5ba1e]
+  - @iqai/adk@0.5.6
+
+## 0.3.27
+
+### Patch Changes
+
+- Updated dependencies [05bb1b8]
+  - @iqai/adk@0.5.5
+
+## 0.3.26
+
+### Patch Changes
+
+- Updated dependencies [2167a47]
+  - @iqai/adk@0.5.4
+
+## 0.3.25
+
+### Patch Changes
+
+- 0082494: fix: env parsing
+
+## 0.3.24
+
+### Patch Changes
+
+- 8143f4f: refactor: simplified agent loaders and managers
+
+## 0.3.23
+
+### Patch Changes
+
+- 737493f: Fix hot reload state sync and State Panel UI issues
+
 ## 0.3.22
 
 ### Patch Changes